pyproject.toml
[project]
name = "yapfm"
version = "0.1.0"
description = ""
authors = [
    {name = "Ephraim Seddor",email = "seddorephraim7@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "tomlkit (>=0.13.3,<0.14.0)",
    "pyyaml (>=6.0.2,<7.0.0)"
]

[tool.poetry]
packages = [{include = "yapfm", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-cov = "^7.0.0"
ruff = "^0.12.12"
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.17.1"
commitizen = "^4.8.4"
pre-commit = "^4.3.0"
types-pyyaml = "^6.0.12.20250822"

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.1.0"
tag_format = "v$version"
update_changelog_on_bump = true
version_files = [
    "pyproject.toml:version",
]

[tool.black]
line-length = 88
target-version = ["py310"]
skip-string-normalization = false
preview = true

[tool.isort]
profile = "black"
line_length = 88

[tool.ruff]
line-length = 88
target-version = "py310"
exclude = ["migrations", ".venv", "venv", "env", ".env", ".bzr", ".direnv", ".eggs", ".git", ".git-rewrite", ".hg", ".ipynb_checkpoints", ".mypy_cache", ".nox", ".pants.d", ".pyenv", ".pytest_cache", ".pytype", ".ruff_cache", ".svn", ".tox", ".venv", ".vscode", "__pypackages__", "_build", "buck-out", "build", "dist", "node_modules", "site-packages"]

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]
fixable = ["ALL"]

[tool.ruff.format]
quote-style = "double"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.10"
disallow_untyped_defs = true
ignore_missing_imports = true
strict_optional = true

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
